Working on refactoring Cosmos::Client
